+++ src/main/java/dynamicquad/agilehub/issue/service/factory/IssueFactory.java
@@ -1,10 +1,10 @@
 package dynamicquad.agilehub.issue.service.factory;
 
-import dynamicquad.agilehub.issue.controller.IssueResponse.AssigneeDto;
-import dynamicquad.agilehub.issue.controller.IssueResponse.ContentDto;
-import dynamicquad.agilehub.issue.controller.IssueResponse.IssueDto;
-import dynamicquad.agilehub.issue.controller.IssueResponse.SubIssueDto;
 import dynamicquad.agilehub.issue.controller.request.IssueRequest.IssueCreateRequest;
+import dynamicquad.agilehub.issue.controller.response.IssueResponse.AssigneeDto;
+import dynamicquad.agilehub.issue.controller.response.IssueResponse.ContentDto;
+import dynamicquad.agilehub.issue.controller.response.IssueResponse.IssueDto;
+import dynamicquad.agilehub.issue.controller.response.IssueResponse.SubIssueDto;
 import dynamicquad.agilehub.issue.domain.Issue;
 import dynamicquad.agilehub.project.domain.Project;
 import java.util.List;
